Fix race condition between main rescheduler task & frame free:
when an async function completes, it does some atomic operations to wake up any `await`ers. when the last async function doing work completes, it wakes up the main async function. There is a race between the last async function checking its awaiters, and the main async function free'ing the last's frame. This addresses that by having the async functions suspend instead of return to avoid checking for an awaiter.